requirements.txt
PyOpenGL
pillow
